Add files via upload
i still need to work on the basement, but here are some change for the first floor.  there is no stairs on the front porch and ramp but i will put them in, on next revision. i just want to get all the big changes out of the way for the first floor.